[333] ETQU pas admin projet, je ne peux pas ajouter, modifier ou supprimer les pages/scripts de mon projet

+++ backend/projects/views.py
@@ -64,6 +64,7 @@ def project_page_list(request, project_uuid):
         serializer = PageSerializer(pages, many=True)
         return JsonResponse(serializer.data, safe=False)
     elif request.method == "POST":
+        check_if_admin_of_project(request.user.id, project.id)
         data = JSONParser().parse(request)
         serializer = PageSerializer(data=data)
         if serializer.is_valid():
@@ -90,6 +91,7 @@ def project_page_detail(request, project_uuid, page_uuid):
         return JsonResponse(serializer.data)
 
     elif request.method == "PUT":
+        check_if_admin_of_project(request.user.id, project.id)
         data = JSONParser().parse(request)
         serializer = PageSerializer(page, data=data)
         if serializer.is_valid():
@@ -98,5 +100,6 @@ def project_page_detail(request, project_uuid, page_uuid):
         return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == "DELETE":
+        check_if_admin_of_project(request.user.id, project.id)
         page.delete()
         return JsonResponse({}, status=status.HTTP_204_NO_CONTENT)
